refactor: Separate send mode config into new file
